Add Preview HTML Functionality and update Egg timer

+++ src/main/java/com/pearson/app/services/ImageService.java
@@ -5,9 +5,12 @@ import com.pearson.app.model.Image;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.List;
 
 
@@ -18,6 +21,9 @@ import java.util.List;
 public class ImageService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ImageService.class);
+
+    @Value("${file.upload.directory}")
+    private String fileUploadDirectory;
 
     @Autowired
     private ImageRepository imageRepository;
@@ -47,14 +53,50 @@ public class ImageService {
         imageRepository.update(image);
     }
 
-    @Transactional
-    public void removeImage(Image image) {
-        imageRepository.delete(image);
+    @Transactional(rollbackFor = Exception.class)
+    public void removeImage(Image image) throws Exception{
+        // delete from filesystem first
+        if(removeImage(image.getNewFilename())) {
+            // then delete from database
+            imageRepository.delete(image);
+        }
+    }
+
+    public boolean removeImage(String filenameStr) throws IOException{
+        LOGGER.debug("Removing Word Document file on filesystem: Filename[{}]", filenameStr);
+
+        boolean deleted = false;
+        File imageFile = new File(fileUploadDirectory+ File.separator + filenameStr);
+        if(imageFile.exists()) {
+            String filenameAbsolutePath = imageFile.getAbsolutePath();
+            LOGGER.debug("Found Word Document file exists on filesystem: File[{}]", filenameAbsolutePath);
+            deleted = imageFile.delete();
+            //File thumbnailFile = new File(fileUploadDirectory+"/"+image.getThumbnailFilename());
+            //thumbnailFile.delete();
+            if (!deleted) {
+                LOGGER.error("Failed to delete file from filesystem: File[{}] Technical Error: [{}]", filenameAbsolutePath, getReasonForFileDeletionFailureInPlainEnglish(imageFile));
+                throw new IOException("Failed to delete file from filesystem: " + filenameAbsolutePath);
+            } else {
+                LOGGER.debug("Successfully deleted file from filesystem: File[{}]", imageFile.getAbsolutePath());
+            }
+        } else {
+            deleted = true;
+        }
+        return deleted;
     }
 
 
-
-
-
+    private String getReasonForFileDeletionFailureInPlainEnglish(File file) {
+        try {
+            if (!file.exists())
+                return "It doesn't exist in the first place.";
+            else if (file.isDirectory() && file.list().length > 0)
+                return "It's a directory and it's not empty.";
+            else
+                return "Somebody else has it open, we don't have write permissions, or somebody stole my disk.";
+        } catch (SecurityException e) {
+            return "We're sandboxed and don't have filesystem access.";
+        }
+    }
 
 }
